% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3BarGraph.R
\name{d3BarGraph}
\alias{d3BarGraph}
\title{D3BarGraph from R}
\usage{
d3BarGraph(df, width = NULL, height = NULL)
}
\arguments{
\item{width}{numeric width for the graph's width in pixels.}

\item{height}{numeric height for the graph's height in pixels.}

\item{barData}{a numerical list of values or a dataframe from R. Used for bar height}
}
\value{
hopefully a graph...and not just a blank screen
}
\description{
Making a bar graph with D3 and R
}

