% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3BarGraph.R
\name{d3BarGraph}
\alias{d3BarGraph}
\title{D3BarGraph from R}
\usage{
d3BarGraph(barData, width = NULL, height = NULL)
}
\arguments{
\item{barData}{a numerical list of values as a dataframe from R. Used for bar height}

\item{width}{numeric width for the network graph's width in pixels.}

\item{height}{numeric height for the graph's height in pixels.}
}
\value{
hopefully a graph...and not just a list in the console
}
\description{
Making a bar graph with D3 and R
}

